Fix size errors.

+++ kernel/include/arch/x86/paging/pae.h
@@ -15,21 +15,21 @@
 // maps 2MB
 struct PAE_PageDirectory_2MB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisabled  :  1;
-    uint32 accessed       :  1;
-    uint32 dirty          :  1;
-    uint32 pageSize       :  1; // must be 1
-    uint32 global         :  1;
-    uint32 ignored        :  3;
-    uint32 pat            :  1;
-    uint32 alwaysZero1    :  8;
-    uint32 address        : 31;
-    uint32 alwaysZero2    : 11;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisabled  :  1;
+    uint64 accessed       :  1;
+    uint64 dirty          :  1;
+    uint64 pageSize       :  1; // must be 1
+    uint64 global         :  1;
+    uint64 ignored        :  3;
+    uint64 pat            :  1;
+    uint64 alwaysZero1    :  8;
+    uint64 address        : 31;
+    uint64 alwaysZero2    : 11;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 1GB
@@ -45,19 +45,19 @@ struct PAE_PageDirectory_2MB
 // maps 4kb
 struct PAE_PageTable_4KB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisabled  :  1;
-    uint32 accessed       :  1;
-    uint32 dirty          :  1;
-    uint32 pat            :  1; // reserved
-    uint32 global         :  1;
-    uint32 ignored        :  3;
-    uint32 address        : 40;
-    uint32 alwaysZero     : 11;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisabled  :  1;
+    uint64 accessed       :  1;
+    uint64 dirty          :  1;
+    uint64 pat            :  1; // reserved
+    uint64 global         :  1;
+    uint64 ignored        :  3;
+    uint64 address        : 40;
+    uint64 alwaysZero     : 11;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 2mb
@@ -69,18 +69,18 @@ struct PAE_PageTable_4KB
 // maps 2mb
 struct PAE_PageDirectory_4KB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisabled  :  1;
-    uint32 accessed       :  1;
-    uint32 ignored1       :  1;
-    uint32 pageSize       :  1; // must be 0
-    uint32 ignored2       :  4;
-    uint32 address        : 40;
-    uint32 alwaysZero     : 11;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisabled  :  1;
+    uint64 accessed       :  1;
+    uint64 ignored1       :  1;
+    uint64 pageSize       :  1; // must be 0
+    uint64 ignored2       :  4;
+    uint64 address        : 40;
+    uint64 alwaysZero     : 11;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 1GB with pages of 4KB
@@ -97,14 +97,14 @@ struct PAE_PageDirectory_4KB
 // maps 1GB. It's same structure for both page sizes (4KB and 2MB).
 struct PAE_PageDirectoryPointerTable_Entry
 {
-    uint32 present       :  1;
-    uint32 alwaysZero1   :  2;
-    uint32 writeThrough  :  1;
-    uint32 cacheDisabled :  1;
-    uint32 alwaysZero2   :  4;
-    uint32 ignored       :  3;
-    uint32 address       : 40;
-    uint32 alwaysZero3   : 10;
+    uint64 present       :  1;
+    uint64 alwaysZero1   :  2;
+    uint64 writeThrough  :  1;
+    uint64 cacheDisabled :  1;
+    uint64 alwaysZero2   :  4;
+    uint64 ignored       :  3;
+    uint64 address       : 40;
+    uint64 alwaysZero3   : 10;
 } __attribute__((packed));
 
 // maps 4GB with pages of 4KB
